+++ client/src/components/Footer/images/vegvesen_logo.png
